Change test case for /statistics/histogram endpoint to fix error
Previously the test case checked whether the length of response
results array was equal to the time span given. However the API returns
results of length 1 if no email is sent during that time, regardless
of the duration of the timespan between starting and ending
timestamps. As a result the test case would fail sometimes and this
fix would prevent it from failing in the above mentioned valid case.